@@ -14,7 +14,7 @@
   ],
   "require": {
     "typo3/cms-core": "^10.4 || ^11 || ^12 || ^13",
-    "php": "^7.2 || ^7.3 || ^7.4 || ^8.0 || ^8.1 || ^8.2 || ^8.3",
+    "php": "^7.4 || ^8.0 || ^8.1 || ^8.2 || ^8.3|| ^8.4",
     "ext-json": "*"
   },
   "require-dev": {
