@@ -69,7 +69,7 @@ class Properties extends AbstractEventHandler implements
      *
      * @var \Brainworxx\Krexx\Service\Factory\Pool
      */
-    protected $pool;
+    protected Pool $pool;
 
     /**
      * Inject the pool.
