@@ -55,7 +55,7 @@ class GetterWithoutGet implements EventHandlerInterface
      *
      * @var Pool
      */
-    protected $pool;
+    protected Pool $pool;
 
     /**
      * {@inheritdoc}
