@@ -46,7 +46,7 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'typo3' => '10.4.0-13.4.99',
-            'php' => '7.2.0-8.3.99',
+            'php' => '7.4.0-8.4.99',
         ],
         'conflicts' => [],
         'suggests' => [
